WAD: Fix byte swapping logic
First, the shifting was by the wrong ammounts, shifting to the left
by 24, then 16, then 8 means you've lost the original bits you had OR-ed
into the new set of bytes, this has been changed to shuffle the bits
along correctly.

Second, shifting to the left on signed integers is undefined
behaviour, so signed implementations have been changed to rely on the
unsigned functions not the other way around.